refactor(platform): split node-executor.ts into focused exec/ modules
node-executor.ts mixed six responsibilities behind one Command object
(process exec, Axios/HTTP driver, WSL named-pipe relay, SSH tunnelling,
request proxying, connection caches) in 1,255 lines.

Split the implementation into single-responsibility modules under
src/platform/exec/ — process-utils, api-driver, commander, ssh-transport,
wsl-relay, proxy-request — and reduce node-executor.ts to a thin facade that
assembles Command and re-exports the identical public surface (named
re-exports only, never export *). Deduplicate the three copies of the stream
bridge into a shared src/utils/streamEmitter.ts.

Pure relocation: the public API of @/platform/node-executor is byte-for-byte
identical, so every import site is untouched; the existing suite is the
regression oracle. check-types, lint, test:run (336/336) and build all pass.
